Move SqlTableSource to tabbedview package.
Also expose class as toplevel import from ftw.tabbedview.
