+++ opengever/meeting/tabs/committeelisting.py
@@ -4,8 +4,8 @@ from ftw.table.interfaces import ITableSourceConfig
 from opengever.meeting import _
 from opengever.meeting.model import Committee
 from opengever.ogds.base.utils import get_current_admin_unit
+from opengever.tabbedview import SqlTableSource
 from opengever.tabbedview.browser.base import BaseListingTab
-from opengever.tabbedview.browser.sqltablelisting import SqlTableSource
 from zope.interface import implements
 from zope.interface import Interface
