+++ opengever/meeting/tabs/submittedproposallisting.py
@@ -3,7 +3,7 @@ from ftw.table.interfaces import ITableSource
 from ftw.table.interfaces import ITableSourceConfig
 from opengever.meeting.model import Proposal
 from opengever.meeting.tabs.proposallisting import ProposalListingTab
-from opengever.tabbedview.browser.sqltablelisting import SqlTableSource
+from opengever.tabbedview import SqlTableSource
 from zope.interface import implements
 from zope.interface import Interface
