+++ opengever/tabbedview/__init__.py
@@ -1,4 +1,5 @@
 from opengever.tabbedview.basesource import GeverTableSource
+from opengever.tabbedview.sqlsource import SqlTableSource
 from zope.i18nmessageid import MessageFactory
 import logging
